Port to game 1.0 + fix three defects the port audit could not see
Game 1.0 (Steam buildid 25007655, after Hovgaard's hotfix reversed the
launch build's Unity/BCL stripping) changed 886 decompiled files against
0.11. This ports the mod onto it.

Permanent 1.0 API changes (20 hunks): AdManager.AdSettings -> top-level
AdSettings; IsBuildingBlockedByAnyService takes an Address; Instantiate-
SingleInstance gained a Transform; RealEstate.purchasePrice is double;
ownerActions is a GameObject[]; EnterBuilding lost 'force'; retailPrice
is an InputField wrapper; ClearLmpCache -> ClearPriceCaches; the diploma
pop-up reads DiplomaData; CharacterInfo uses totalBusinessesLabel;
SellBuildingForCompat takes double. Sad-period patch deleted (1.0 removed
the whole subsystem). Two dead tick-wall entries removed, three new 1.0
hourly steps added. Save carry-forward isolates the native version walk
behind a NoInlining wrapper. Protocol 18 -> 19.

Three defects the audit could not see, because all three are invisible to
a compiler, to a member-reference gate, and to Harmony binding:

  BizPhone icons and height (compiler-invisible STRING). Injection walked
  a fixed scene path, "Container/Phone/AppButtons". 1.0 rebuilt the phone
  around the new SmartPhoneFrame (equippable cases that also rewrite the
  body rect), so both app buttons and the extra row vanished - announced
  only by 53 repetitions of a warning in a log nobody grepped. Now binds
  to SmartphoneUI.appButtonTemplate, a serialized field identical in both
  versions, and takes its parent; the old path stays as a fallback so this
  can only degrade to the previous behaviour. Field-confirmed fixed.

  Client taxis (NEW GAME CODE calling into objects we fake). 1.0 added a
  tail to TaxiSystem.TravelCoroutine that re-spawns the cab, via a new
  TaxiController.GetVehiclePrefab. On a ghost that throws - and the two
  lines after it are UiFader.UnFade() and _travelCoroutine = null, while
  TravelTo early-returns whenever _travelCoroutine is non-null. So the
  client kept a faded screen AND every later ride that session was
  silently refused. Substitutes a pooled taxi prefab only when the native
  path would throw; contains InstantiateVehicle so the coroutine always
  finishes. Field-confirmed fixed - and the probe corrected the diagnosis:
  _vehicleComponent is present, its .prefab is null.

  Tuition tax bucket (compiler-invisible STRING). 1.0 renamed it to
  "ba:businesstype_school"; the old value filed the instant-diploma
  tuition into an orphan deduction bucket, silently.

Also reverts three client suppressions added earlier the same day. They
were wrong: all three ticks operate on per-machine save state the mod
never syncs (saves are per-player), so suppressing prevented no double-run
and instead stranded the client's own food deliveries - blocking every
later order - emptied their gig board, and disabled their pricing manager.
The reasoning is kept in place as a comment rather than deleted, because
the next reader will have the same idea. The real hazard was the reverse:
a pricing plan sweeps its whole NEIGHBOURHOOD for anything flagged rented-
by-player, and another player's shop carries that flag on every machine.
Handled by a merger-veil entry plus a foreign-business filter; a grant is
not required for the hazard, so permissions alone would not have covered it.

Adds probe P-RESTOCK-AUTHORITY, which turns "stand in a shop and watch the
shelves for an in-game hour" into one log line. It reads ownership at
Priority.Last, after the veil un-flips, because that is the value the game
actually uses.

Field-verified on two machines: 390 classes OK, 0 failed, 0 dead, and zero
native exceptions in either log. Three configs build clean; unit suite 90/90.

Protocol 18 -> 19 means every player must update together.

+++ src/MergerWallet.cs
@@ -77,7 +77,11 @@ namespace BigAmbitionsMP
         // overloads — the name-only patch failed "Ambiguous match" and the wallet bypass guard was
         // silently NOT applied. The (float, Address) overload is the one that writes Money directly;
         // the other two funnel into it.
-        [HarmonyPatch(typeof(Helpers.RealEstateHelper), "SellBuildingForCompat", typeof(float), typeof(Address))]
+        // 1.0 retyped the price parameter float -> double (2026-08-29). The old typeof(float) still
+        // BOUND, because .NET's default binder widens float->double when only one candidate matches —
+        // but that is leniency, not a contract, and would turn ambiguous the moment another overload
+        // appears. Declared exactly now. The pre/postfix only snapshot Money, so behaviour is unchanged.
+        [HarmonyPatch(typeof(Helpers.RealEstateHelper), "SellBuildingForCompat", typeof(double), typeof(Address))]
         public static class Patch_RealEstateHelper_SellForCompat_WalletForward
         {
             static void Prefix(out float __state) => __state = MoneyNow();
